+++ packages/utils/src/TicketUtils.ts
@@ -1,7 +1,13 @@
-import { Column, RegularGame, Ticket } from '@caglarturali/piyango-common';
+import {
+  Column,
+  RegularDrawData,
+  RegularGame,
+  Ticket,
+} from '@caglarturali/piyango-common';
+import { RegularCheck } from './Check';
 
 export class TicketUtils {
-  private readonly letters = 'ABCDEFGH';
+  private static readonly letters = 'ABCDEFGH';
   private game: RegularGame;
   private cols: Column[];
 
@@ -25,21 +31,65 @@ export class TicketUtils {
       (chunk) =>
         ({
           game: this.game,
-          numbers: this.arrayToObjectMap(chunk),
+          numbers: TicketUtils.arrayToObjectMap(this.game, chunk),
         } as Ticket),
     );
   }
 
   /**
+   * Compares tickets against given draw data.
+   * @param drawData Regular draw data
+   */
+  compareAgainstDraw(drawData: RegularDrawData) {
+    const tickets = this.tickets();
+    return tickets.map((ticket) =>
+      TicketUtils.compareTicketAgainstDraw(ticket, drawData),
+    );
+  }
+
+  /**
+   * Compares single ticket against given draw data.
+   * @param ticket Ticket
+   * @param drawData Regular draw data
+   */
+  static compareTicketAgainstDraw(ticket: Ticket, drawData: RegularDrawData) {
+    const { game, numbers } = ticket;
+    const cols = TicketUtils.objectToArrayMap(numbers);
+    const check = RegularCheck.fromColumns(game as RegularGame, drawData, cols);
+    check.validate();
+    check.process();
+
+    return {
+      ticket,
+      results: check.results,
+    };
+  }
+
+  /**
    * Converts Columns array into string indexed object.
+   * @param game Regular game object
    * @param arr Array of Column objects
    */
-  private arrayToObjectMap(arr: Column[]) {
+  static arrayToObjectMap(game: RegularGame, arr: Column[]) {
     const obj: { [key: string]: Column } = {};
     arr.forEach((col, i) => {
-      const colName = this.letters[i % this.game.columns];
+      const colName = TicketUtils.letters[i % game.columns];
       obj[colName] = col;
     });
     return obj;
+  }
+
+  /**
+   * Converts string indexed object into Columns array.
+   * @param obj Object of Columns
+   */
+  static objectToArrayMap(obj: { [colName: string]: Column }) {
+    const cols: Column[] = [];
+    for (const colName in obj) {
+      if (obj.hasOwnProperty(colName)) {
+        cols.push(obj[colName]);
+      }
+    }
+    return cols;
   }
 }