+++ packages/utils/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2020-04-07
+
+### Added
+
+- RegularCheck.checkMatch static method
+- TicketUtils.compareTicketAgainstDraw static method
+- TicketUtils.arrayToObjectMap static method
+- TicketUtils.objectToArrayMap static method
+
 ## [1.2.0] - 2020-04-07
 
 ### Removed